+++ FastAPI/Model/User.py
@@ -5,10 +5,13 @@ from pydantic import BaseModel
 
 class User(BaseModel):
     username: str
-
     full_name: Optional[str] = None
     email: str = None
     disabled: Optional[bool] = None
+
+
+class User_Create(User):
+    password: int
 
 
 class User_in_DB(User):
