+++ Security.py
@@ -8,7 +8,7 @@ from starlette import status
 from FastAPI.Model.Token import TokenData
 from FastAPI.Model.User import User
 from FastAPI.Usecases.user_operations import user_op_get_user_data
-from main import logger
+from normal_main import logger
 
 SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
 ALGORITHM = "HS256"
@@ -50,7 +50,7 @@ def security_op_security_check(token: str):
 
     user: User = user_op_get_user_data(token_data.user_name)
     if not user:
-        raise  HTTPException(status_code=400, detail="Wrong user.")
+        raise HTTPException(status_code=400, detail="Wrong user.")
     user_data = User(**user)
 
     logger.debug("User here is : {}".format(user))
@@ -60,4 +60,4 @@ def security_op_security_check(token: str):
     if user_data.disabled:
         raise HTTPException(status_code=400, detail="User is disabled.")
 
-    return user
+    return user_data
