+++ docker-compose.yml
@@ -1,0 +1,10 @@
+version: '3.8'
+
+services:
+  web:
+    build: ./
+    command: uvicorn app.main:app --reload --workers 1 --host 0.0.0.0 --port 8000
+    volumes:
+      - ./src/:/usr/src/app/
+    ports:
+      - 8002:8000
